refactor(dashboard-core): require injected folder navigation
