+++ packages/dashboard-core/test/unit/declaredDependencies.test.ts
@@ -5,6 +5,7 @@ import { describe, expect, it } from "vitest";
 
 const PACKAGE_ROOT = join(dirname(fileURLToPath(import.meta.url)), "..", "..");
 const NODE_BUILTIN_PREFIX = "node:";
+const PROHIBITED_NODE_IMPORTS = new Set(["node:fs", "node:fs/promises", "node:os", "node:path"]);
 
 function packageNameOf(specifier: string): string | undefined {
   if (specifier.startsWith(".") || specifier.startsWith(NODE_BUILTIN_PREFIX)) return undefined;
@@ -12,17 +13,25 @@ function packageNameOf(specifier: string): string | undefined {
   return specifier.startsWith("@") ? segments.slice(0, 2).join("/") : segments[0];
 }
 
-function collectSourceImports(directory: string, found: Map<string, string[]>): void {
+function collectSourceImports(
+  directory: string,
+  found: Map<string, string[]>,
+  prohibitedNodeImports: string[],
+): void {
   for (const entry of readdirSync(directory)) {
     const full = join(directory, entry);
     if (statSync(full).isDirectory()) {
-      collectSourceImports(full, found);
+      collectSourceImports(full, found, prohibitedNodeImports);
       continue;
     }
     if (!entry.endsWith(".ts")) continue;
     const text = readFileSync(full, "utf8");
     for (const match of text.matchAll(/from\s*"([^"]+)"/g)) {
-      const name = packageNameOf(match[1]);
+      const specifier = match[1];
+      if (specifier !== undefined && PROHIBITED_NODE_IMPORTS.has(specifier)) {
+        prohibitedNodeImports.push(`${specifier} <- ${full.slice(PACKAGE_ROOT.length + 1)}`);
+      }
+      const name = specifier === undefined ? undefined : packageNameOf(specifier);
       if (name === undefined) continue;
       const list = found.get(name) ?? [];
       list.push(full.slice(PACKAGE_ROOT.length + 1));
@@ -38,7 +47,8 @@ describe("dashboard-core declared dependencies", () => {
     };
     const declared = new Set(Object.keys(manifest.dependencies ?? {}));
     const imports = new Map<string, string[]>();
-    collectSourceImports(join(PACKAGE_ROOT, "src"), imports);
+    const prohibitedNodeImports: string[] = [];
+    collectSourceImports(join(PACKAGE_ROOT, "src"), imports, prohibitedNodeImports);
 
     const undeclared = [...imports.entries()]
       .filter(([name]) => !declared.has(name))
@@ -47,5 +57,6 @@ describe("dashboard-core declared dependencies", () => {
     // Emitted declarations re-export src type imports, so an undeclared src
     // dependency becomes an undeclared production type dependency for consumers.
     expect(undeclared).toEqual([]);
+    expect(prohibitedNodeImports).toEqual([]);
   });
 });
