+++ station/src/station/store/stationCommandDispatch.test.ts
@@ -14,6 +14,7 @@ import { dashboardRowIds, selectDashboardSlots } from "@station/dashboard-core/s
 import { createObserverStationClient } from "../../sources/observerStationClient.js";
 import type { StationClient } from "../../sources/types.js";
 import { waitFor } from "../../terminal/testing/waitFor.js";
+import { createFakeFolderService } from "../test/support/fakeFolderService.js";
 import type { StationMouseEvent } from "../../input/mouse.js";
 import { externalAgentSnapshot, manyProjectsSnapshot } from "../fixtures/scenarios.js";
 import { routeStationMouse } from "../input/stationMouse.js";
@@ -67,7 +68,9 @@ describe("station command dispatch through the shared client", () => {
         exitRenderer: () => dashboardExecution({ kind: "success" }),
       },
     };
-    const store = createStationDashboardRuntime(client, capabilities);
+    const store = createStationDashboardRuntime(client, capabilities, {
+      folderService: createFakeFolderService(),
+    });
     store.start();
     client.start();
     const harness: Harness = { fake, client, store, detach: () => store.dispose() };
